% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_matrix.R
\name{.epsilon_matrix}
\alias{.epsilon_matrix}
\title{.epsilon_matrix}
\usage{
.epsilon_matrix(Z, nperm, surv, factor_matrix, covariates = NULL, dat)
}
\arguments{
\item{Z}{the score vector returned by \code{.survival_scores()} function.}

\item{nperm}{number of permutations performed to estimate the \code{epsilon} matrix.
Default is \code{300}.}

\item{surv}{a \code{\link[survival]{Surv}} object of length \code{n}}

\item{factor_matrix}{a \code{n x p} \code{data.frame} of the expression for the
particular gene set of interest being tested}

\item{covariates}{a \code{n x l} matrix of the covariates to adjust upon. Default is \code{NULL}}

\item{dat}{data used to fit survival model returned by \code{.survival_scores()} function.}
}
\value{
The epsilon matrix.
}
\description{
Compute the epsilon matrix by permutation for the \code{sGBJ_scores()} function.
}
