% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_scores.R
\name{.survival_scores}
\alias{.survival_scores}
\title{.survival_scores}
\usage{
.survival_scores(counts_pathway, covariates = NULL, surv)
}
\arguments{
\item{counts_pathway}{a data frame of the counts for the particular pathway of interest of size nxp}

\item{covariates}{a matrix nxl of the covariates to adjust (default=NULL)}

\item{surv}{a surv object of size n}
}
\value{
A list of length 2 with the updated counts_pathway (same as counts_pathway but removing pathways for which survival model failed to converge) and the Z matrix.
}
\description{
Compute survival score.
}
