% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_matrix.R
\name{.epsilon_matrix}
\alias{.epsilon_matrix}
\title{.epsilon_matrix}
\usage{
.epsilon_matrix(Z, nperm, surv, factor_matrix, covariates = NULL, datas)
}
\arguments{
\item{Z}{The score vector returned by .survival_scores() function.}

\item{nperm}{number of permutations to perform to estimate the matrix epsilon (default=300)}

\item{surv}{a surv object of size n}

\item{factor_matrix}{a data frame of the counts for the particular pathway of interest of size nxp}

\item{covariates}{a matrix nxl of the covariates to adjust (default=NULL)}

\item{datas}{Data used to fit survival model returned by .survival_scores() function.}
}
\value{
The epsilon matrix.
}
\description{
Compute the epsilon matrix by permutation for the sGBJ_scores() function.
}
