% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_scores.R
\name{.survival_scores}
\alias{.survival_scores}
\title{.survival_scores}
\usage{
.survival_scores(factor_matrix, covariates = NULL, surv)
}
\arguments{
\item{factor_matrix}{a \code{n x p} \code{data.frame} of the expression for
the particular gene set of interest being tested}

\item{covariates}{a matrix nxl of the covariates to adjust. Default is \code{NULL}}

\item{surv}{a \code{\link[survival]{Surv}} object of length \code{n}}
}
\value{
A list of length 3 with the updated factor_matrix (same as factor_matrix but removing columns for which survival model failed to converge), the Z matrix and the data used to fit survival model.
}
\description{
Compute the survival score
}
