% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_scores.R
\name{.survival_scores}
\alias{.survival_scores}
\title{.survival_scores}
\usage{
.survival_scores(factor_matrix, covariates = NULL, surv)
}
\arguments{
\item{factor_matrix}{a data frame of the counts for the particular pathway of interest of size nxp}

\item{covariates}{a matrix nxl of the covariates to adjust (default=NULL)}

\item{surv}{a surv object of size n}
}
\value{
A list of length 3 with the updated factor_matrix (same as factor_matrix but removing columns for which survival model failed to converge), the Z matrix and the data used to fit survival model.
}
\description{
Compute survival score.
}
