% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGBJ.R
\name{sGBJ}
\alias{sGBJ}
\title{Compute the pvalues and GBJ value associated with a pathway and survival}
\usage{
sGBJ(surv, counts_pathway, covariates = NULL)
}
\arguments{
\item{surv}{a surv object of size n}

\item{counts_pathway}{a data frame of the counts for the particular pathway of interest of size nxp}

\item{covariates}{a matrix nxl of the covariates to adjust (default=NULL)}
}
\value{
The GBJ value and it's pvalue associated
}
\description{
Compute the pvalues and GBJ value associated with a pathway and survival
}
\examples{
# sGBJ(surv,counts_pathway)
}
