% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGBJ_scores.R
\name{sGBJ_scores}
\alias{sGBJ_scores}
\title{Compute the pvalues and GBJ value associated with a pathway and survival}
\usage{
sGBJ_scores(surv, counts_pathway, covariates = NULL, nperm = 300)
}
\arguments{
\item{surv}{a surv object of size n}

\item{counts_pathway}{a data frame of the counts for the particular pathway of interest of size nxp}

\item{covariates}{a matrix nxl of the covariates to adjust (default=NULL)}

\item{nperm}{number of permutations to perform to estimate the matrix epsilon (default=300)}
}
\value{
The GBJ value and it's pvalue associated
}
\description{
Compute the pvalues and GBJ value associated with a pathway and survival
}
\examples{
 n <- 5
 surv_data <- data.frame(Time = runif(n = n, min = 0, max = 100),
                         event = rbinom(n = n, size = 1, prob = 0.5))
 surv <- survival::Surv(time = surv_data$Time, event = surv_data$event)

 counts_pathway <- data.frame(P1 = rnorm(n = n),
                              P2 = rnorm(n = n))

 sGBJ::sGBJ_scores(surv,counts_pathway, nperm = 2)

 # with covariates

 covariates <- data.frame(age = runif(n = n, 60, 90))

 sGBJ_scores(surv,counts_pathway, nperm = 2, covariates = covariates)
}
